requirements.txt
accelerate
argh
dacite
demucs
diffusers>=0.9.0
flask
flask_cors
numpy
pillow>=9.1.0
plotly
pydub
pysoundfile
scipy
soundfile
sox
streamlit>=1.10.0
torch
torchaudio
torchvision
transformers
git+https://github.com/cloneofsimo/lora.git
